LayoutTests/http/tests/inspector/console/message-from-iframe.html

<script src="/inspector/resources/inspector-test.js"></script>
<script>
    function test() {
        // The purpose of these tests are to test that Web Inspector should receive console messages generated from various iframe tree correctly.

        const suite = InspectorTest.createAsyncSuite("MessageFromIFrame");

        // Utilities for test code.
        let callback = () => {};
        WI.consoleManager.addEventListener(WI.ConsoleManager.Event.MessageAdded, (event) => {
            InspectorTest.log(`Got message: ${event.data.message._messageText}`);
            callback();
        });

        function runMessageTest(resolve, expected, func) {
            let seen = 0;

            callback = () => {
                if (++seen === expected)
                    resolve();
            };

            InspectorTest.evaluateInPage(func);
        }
        // Utilities for page code.
        InspectorTest.evaluateInPage(`
            function addIFrame(url) {
                const iframe = document.createElement("iframe");
                iframe.src = url;
                iframe.onload = () => console.log("iframe is loaded in " + location.href);
                document.body.appendChild(iframe);
            }
        `);

        suite.addTestCase({
            name: "MessageFromIFrame.SameOriginIFrame",
            description: "console message comming from same origin iframe",
            test: (resolve) => runMessageTest(resolve, 4, `addIFrame("resources/console-messages.html")`),
        });

        suite.addTestCase({
            name: "MessageFromIFrame.CrossOriginIFrame",
            description: "console message comming from cross origin iframe",
            test: (resolve) => runMessageTest(resolve, 4, `addIFrame("http://localhost:8000/inspector/console/resources/console-messages.html")`),
        });

        suite.addTestCase({
            name: "MessageFromIFrame.GrandChildSameOriginIFrameInSameOrigin",
            description: "console message comming from same origin grand-child iframe in same origin iframe",
            test: (resolve) => runMessageTest(resolve, 5, `addIFrame("resources/embedded-same-origin.html")`),
        });

        suite.addTestCase({
            name: "MessageFromIFrame.GrandChildSameOriginIFrameInCrossOrigin",
            description: "console message comming from same origin grand-child iframe in cross origin iframe",
            test: (resolve) => runMessageTest(resolve, 5, `addIFrame("http://localhost:8000/inspector/console/resources/embedded-cross-origin.html")`),
        });

        suite.runTestCasesAndFinish();
    }
</script>

<body onload="runTest()"></body>
